refactor: 도메인 패키지 위치조정
