+++ src/main/java/com/example/hotdeal/domain/product/product/domain/Product.java
@@ -7,7 +7,6 @@ import lombok.Setter;
 
 @Entity
 @Getter
-@Setter
 @Table(name = "products")
 public class Product extends BaseEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
